Projeto API-Produto com prometheus e grafana

+++ src/main/java/br/com/demoapi/controller/ProdutoController.java
@@ -3,13 +3,13 @@ package br.com.demoapi.controller;
 import br.com.demoapi.entity.Produto;
 import br.com.demoapi.repository.ProdutoRepository;
 import br.com.demoapi.service.ProdutoService;
-import org.springdoc.api.OpenApiResourceNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/produto")
@@ -28,7 +28,13 @@ public class ProdutoController {
     public ResponseEntity<List<Produto>> buscarTodos() {
         return ResponseEntity.ok(produtoService.buscarTodos());
     }
-
+    @GetMapping("/{id}")
+    public ResponseEntity<Optional<Produto>> buscar(@PathVariable Long id) {
+        if (!produtoRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(produtoService.buscarPorId(id));
+    }
     @PostMapping
     public ResponseEntity<Produto> salvar(@RequestBody Produto produto) {
         var novoProduto = produtoService.salvar(produto);
@@ -37,17 +43,20 @@ public class ProdutoController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> remover(@PathVariable Long id) {
+        if (!produtoRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         produtoService.delete(id);
         return ResponseEntity.ok().build();
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<Produto> atualizar(@PathVariable Long id, @RequestBody Produto produto) {
-        var produtoOptional = produtoRepository.findById(id);
-        if (produtoOptional.isEmpty()){
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        if (!produtoRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
         }
         produto.setId(id);
-        return ResponseEntity.ok(produtoRepository.save(produto));
+        Produto produtoatualizado = produtoService.salvar(produto);
+        return ResponseEntity.ok(produtoatualizado);
     }
 }
